Add SudokuFrame Class
Added the SudokuFrame class which will hold the Sudoku Puzzle in the form of a 2D matrix. This class will also take in values and store them into the matrix. The object of this class can be used in the SudokuSolver object. It will be easier to manipulate and display values.