Create example.cpp, Rename sudoku.cpp -> sudoku.h
Renamed the earlier sudoku.cpp file to sudoku.h so that it can be used as a header file in other programs. Also, created a file called example.cpp which tells the user how the header file can be included in the program and how the SudokuSolver object can be called.